+++ WadGraphEs.MetricsEndpoint/WadGraphEs.MetricsEndpoint/ApiControllers/StatusController.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+using System.Web.Http;
+
+namespace WadGraphEs.MetricsEndpoint.ApiControllers {
+	[AllowAnonymous]
+	public class StatusController : ApiController{
+		public string Get() {
+			return "Running";
+		}
+	}
+}
