+++ WadGraphEs.MetricsEndpoint/WadGraphEs.MetricsEndpoint/App_Start/WebRoutes.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+using System.Web.Routing;
+using System.Web.Mvc;
+
+namespace WadGraphEs.MetricsEndpoint {
+    public class WebRoutes {
+        internal static void Register(System.Web.Routing.RouteCollection routes) {
+            routes.MapRoute("home", "", new { controller = "Home", action = "Index" });
+            routes.MapRoute("login", "login", new { controller = "Account", action = "Login" });
+        }
+    }
+}
