+++ WadGraphEs.MetricsEndpoint/WadGraphEs.MetricsEndpoint/Global.asax.cs
@@ -13,6 +13,7 @@ namespace WadGraphEs.MetricsEndpoint {
     public class MvcApplication : System.Web.HttpApplication {
         protected void Application_Start() {
             WebApiConfig.Register(GlobalConfiguration.Configuration);
+            WebRoutes.Register(RouteTable.Routes);
         }
     }
 }
