+++ WadGraphEs.MetricsEndpoint/WadGraphEs.MetricsEndpoint/MVC/Controllers/HomeController.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+using System.Web.Mvc;
+
+namespace WadGraphEs.MetricsEndpoint.MVC.Controllers {
+    public class HomeController : Controller{
+        public ActionResult Index() {
+            return Content("ok");
+        }
+    }
+}
